+++ tests/test_yaml.py
@@ -1,6 +1,7 @@
 import unittest
-from troposphere import Template, Output, Ref
+from troposphere import Template, Output, Ref, Sub
 from troposphere import s3
+from troposphere.cloudformation import WaitCondition
 
 
 s3_bucket_yaml = """\
@@ -14,6 +15,38 @@ Resources:
     Properties:
       AccessControl: PublicRead
     Type: AWS::S3::Bucket
+"""
+
+cond_string = """My
+special multiline
+Handle"""
+
+cond_normal = """Resources:
+  MyWaitCondition:
+    Properties:
+      Handle: !Sub "My\\nspecial multiline\\nHandle"
+      Timeout: 30
+    Type: AWS::CloudFormation::WaitCondition
+"""
+
+cond_long = """Resources:
+  MyWaitCondition:
+    Properties:
+      Handle:
+        Fn::Sub: "My\\nspecial multiline\\nHandle"
+      Timeout: 30
+    Type: AWS::CloudFormation::WaitCondition
+"""
+
+cond_clean = """Resources:
+  MyWaitCondition:
+    Properties:
+      Handle: !Sub |-
+        My
+        special multiline
+        Handle
+      Timeout: 30
+    Type: AWS::CloudFormation::WaitCondition
 """
 
 
@@ -30,3 +63,13 @@ class TestYAML(unittest.TestCase):
             Description="Name of S3 bucket to hold website content"
         ))
         self.assertEqual(s3_bucket_yaml, t.to_yaml())
+
+    def test_yaml_long_form(self):
+        t = Template()
+        t.add_resource(WaitCondition(
+            "MyWaitCondition", Timeout=30, Handle=Sub(cond_string)))
+        self.assertEqual(cond_normal, t.to_yaml())
+        self.assertEqual(cond_long, t.to_yaml(long_form=True))
+        self.assertEqual(cond_long, t.to_yaml(False, True))
+        self.assertEqual(cond_clean, t.to_yaml(clean_up=True))
+        self.assertEqual(cond_clean, t.to_yaml(True))