+++ troposphere/elasticsearch.py
@@ -49,6 +49,13 @@ class ElasticsearchClusterConfig(AWSProperty):
     }
 
 
+class EncryptionAtRestOptions(AWSProperty):
+    props = {
+        'Enabled': (boolean, False),
+        'KmsKeyId': (basestring, False),
+    }
+
+
 class SnapshotOptions(AWSProperty):
     props = {
         'AutomatedSnapshotStartHour': (integer_range(0, 23), False)
@@ -72,6 +79,7 @@ class Domain(AWSObject):
         'EBSOptions': (EBSOptions, False),
         'ElasticsearchClusterConfig': (ElasticsearchClusterConfig, False),
         'ElasticsearchVersion': (basestring, False),
+        'EncryptionAtRestOptions': (EncryptionAtRestOptions, False),
         'SnapshotOptions': (SnapshotOptions, False),
         'Tags': ((Tags, list), False),
         'VPCOptions': (VPCOptions, False)